requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
ultralytics>=8.0.0
aiohttp>=3.9.0
pillow>=10.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
googletrans==4.0.0rc1
numpy>=1.26.0
opencv-python>=4.8.0
